Moving type to its own file
