+++ src/Components/Aspire.Client/ConfigurationSchemaAttributes.cs
@@ -37,24 +37,3 @@ public sealed class ConfigurationSchemaAttribute : Attribute
     /// </summary>
     public string[]? ExclusionPaths { get; }
 }
-
-/// <summary>
-/// Provides information to describe the logging categories produced by a component.
-/// </summary>
-[AttributeUsage(AttributeTargets.Assembly, AllowMultiple = false)]
-public sealed class LoggingCategoriesAttribute : Attribute
-{
-    /// <summary>
-    /// Initializes a new instance of the <see cref="LoggingCategoriesAttribute"/> class.
-    /// </summary>
-    /// <param name="categories">The list of log categories produced by the component. These categories will show up under the Logging:LogLevel section in appsettings.json.</param>
-    public LoggingCategoriesAttribute(params string[] categories)
-    {
-        Categories = categories;
-    }
-
-    /// <summary>
-    /// The list of log categories produced by the component. These categories will show up under the Logging:LogLevel section in appsettings.json.
-    /// </summary>
-    public string[] Categories { get; set; }
-}
